copula.trans {regpro}	R Documentation

Makes a copula transformation

Description

Transforms a data matrix so that the marginals follow approximately the standard Gaussian distribution. Alternatively, the marginals can be transformed to follow approximately the uniform distribution on [0,1].

Usage

copula.trans(dendat, marginal=rep("gauss",dim(dendat)[2]), remna=TRUE)

Arguments

dendat
n*d data matrix; the data matrix of n observations and d variables

marginal
d-vector of character strings; the character strings can be "gauss" or "uniform"

remna
TRUE or FALSE; if remna=TRUE, then the rows containing a NA are removed

Value

n*d data matrix

Author(s)

Jussi Klemela

Examples

set.seed(2)
n<-100
d<-2 
dendat<-matrix(runif(n*d),n,d)
x<-copula.trans(dendat)

[Package regpro version 0.1.1 Index]