emp.distribu {regpro}	R Documentation

Empirical distribution function at one point

Description

Computes the value of an empirical distribution function at one point.

Usage

emp.distribu(arg, dendat)

Arguments

arg
d-vector; the point where the estimate is evaluated

dendat
n*d data matrix; the data matrix of n observations and d variables

Value

a real number or a d vector; if d>1 the empirical distribution function is estimated for each column of the data matrix "dendat"

Author(s)

Jussi Klemela

See Also

emp.quantile,

Examples

set.seed(2)
n<-100
d<-2 
x<-matrix(runif(n*d),n,d)

arg<-c(0,0)
emp.distribu(arg,x)

[Package regpro version 0.1.1 Index]