linear.quan {regpro}	R Documentation

Multivariate linear quantile regression estimator

Description

Computes the estimates of parameters for a linear quantile regression estimator.

Usage

linear.quan(x, y, p=0.5)

Arguments

x
n*d data matrix; the matrix of the values of the explanatory variables

y
n vector; the values of the response variable

p
0<p<1; the p:th conditional quantile function will be estimated

Details

numerical optimization is used in the calculation

Value

list of beta0 and beta1; beta0 is a real number and beta1 is a d vector; beta0 is the estimate of the intercept and beta1 is the vector containing the estimates of the coefficients

Author(s)

Jussi Klemela

See Also

linear,

Examples

set.seed(1)
n<-100
d<-2 
x<-8*matrix(runif(n*d),n,d)-3
C<-(2*pi)^(-d/2)
phi<-function(x){ return( C*exp(-sum(x^2)/2) ) }
D<-3; c1<-c(0,0); c2<-D*c(1,0); c3<-D*c(1/2,sqrt(3)/2)
func<-function(x){phi(x-c1)+phi(x-c2)+phi(x-c3)}
y<-matrix(0,n,1)
for (i in 1:n) y[i]<-func(x[i,])+0.01*rnorm(1)

linear.quan(x,y)

[Package regpro version 0.1.1 Index]