ma {regpro}	R Documentation

Moving average of a time series

Description

Computes a one-sided weighted moving average from a univariate time series. The one-sided moving average can be used to predict the next value of the sequence.

Usage

ma(x, h=1, kernel="exp", k=length(x))

Arguments

x
n vector; the observed values of the time series

h
a positive real number; the smoothing parameter of the moving average

kernel
a character; determines the kernel function; either "exp", "uniform", "gauss", or "bart"

k
a positive integer; the moving average includes at most k observations

Value

a real number

Author(s)

Jussi Klemela

See Also

kernesti.regr,

Examples

set.seed(1)
n<-100
x<-runif(n)
ma(x)

[Package regpro version 0.1.1 Index]